fix(runner): keep acceptance effects safe for browser imports
